+++ appinfo/routes.php
@@ -9,18 +9,17 @@
  * @copyright Jörn Friedrich Dreyer 2015
  */
 
-namespace OCA\Impersonate\AppInfo;
-
-/**
- * Create your routes in here. The name is the lowercase name of the controller
- * without the controller part, the stuff after the hash is the method.
- * e.g. page#index -> PageController->index()
- *
- * The controller class has to be registered in the application.php file since
- * it's instantiated in there
- */
-$application = new Application();
-
-$application->registerRoutes($this, array('routes' => array(
-	array('name' => 'settings#impersonate', 'url' => '/user', 'verb' => 'POST'),
-)));
+use OCP\AppFramework\App;
+$application = new App('impersonate');
+$application->registerRoutes(
+	$this,
+	[
+		'routes' => [
+			[
+				'name' => 'Settings#impersonate',
+				'url' => '/user',
+				'verb' => 'POST',
+			],
+		],
+	]
+);