+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+owncloud-impersonate (0.0.3)
+* use eventdispatcher for js injection to users page, requires oc9
+
+owncloud-impersonate (0.0.2)
+* add action to userlist instead of input field in admin settings
+
 owncloud-impersonate (0.0.1)
 * First release